requirements.txt
tqdm
requests
argparse
Shimmy==2.0.0
gymnasium==1.1.1
dm_control==1.0.28
imageio==2.37.0
numpy==2.2.5
matplotlib
torch==2.6.0